Design Document
A document that goes in-depth on current competition, narrowed down on a specific region and where this system would be most effective. Used schematics and 3D visual designs to bring the system to life. Created a BOM and an installation manual. Showed the possible risks involved in the system. Finally, outlined a 6 test evaluation that will be implemented to see if system can deliver on the requirements.